requirements.txt
-i https://pypi.org/simple
alabaster==0.7.16; python_version >= '3.9'
amqp==5.2.0; python_version >= '3.6'
asgiref==3.7.2; python_version >= '3.7'
async-timeout==4.0.3; python_full_version <= '3.11.2'
attrs==23.2.0; python_version >= '3.7'
babel==2.14.0; python_version >= '3.7'
beautifulsoup4==4.12.3; python_full_version >= '3.6.0'
billiard==4.2.0; python_version >= '3.7'
celery[redis]==5.3.6
certifi==2024.2.2; python_version >= '3.6'
cfgv==3.4.0; python_version >= '3.8'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
click-didyoumean==0.3.0; python_full_version >= '3.6.2' and python_full_version < '4.0.0'
click-plugins==1.1.1
click-repl==0.3.0; python_version >= '3.6'
colorama==0.4.6; platform_system == 'Windows'
coverage==7.4.1
crispy-bootstrap5==2023.10
cron-descriptor==1.4.3
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
diff-match-patch==20230430; python_version >= '3.7'
distlib==0.3.8
django==4.2.10
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-coverage-plugin==3.1.0
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-docs==0.3.3
django-extensions==3.2.3
django-extensions-models==1.0
django-filter==23.5
django-health-check==3.18.1
django-import-export==3.3.7
django-redis==5.4.0
django-stubs==4.2.7
django-stubs-ext==4.2.7; python_version >= '3.8'
django-timezone-field==6.1.0; python_version >= '3.8' and python_version < '4.0'
djangorestframework==3.14.0
docutils==0.20.1; python_version >= '3.7'
drf-spectacular==0.27.1
et-xmlfile==1.1.0; python_version >= '3.6'
factory-boy==3.3.0
faker==23.1.0; python_version >= '3.8'
filelock==3.13.1; python_version >= '3.8'
flower==2.0.1
humanize==4.9.0; python_version >= '3.8'
identify==2.5.34; python_version >= '3.8'
idna==3.6; python_version >= '3.5'
imagesize==1.4.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
inflection==0.5.1; python_version >= '3.5'
jinja2==3.1.3; python_version >= '3.7'
jsonschema==4.21.1; python_version >= '3.8'
jsonschema-specifications==2023.12.1; python_version >= '3.8'
kombu==5.3.5; python_version >= '3.8'
linkpreview==0.8.3
markdown==3.5.2
markuppy==1.14
markupsafe==2.1.5; python_version >= '3.7'
mypy==1.8.0
mypy-extensions==1.0.0; python_version >= '3.5'
mysqlclient==2.2.4
nodeenv==1.8.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2; python_version >= '3.7'
pillow==10.2.0
platformdirs==4.2.0; python_version >= '3.8'
pprintpp==0.4.0
pre-commit==3.6.1
prometheus-client==0.19.0; python_version >= '3.8'
prompt-toolkit==3.0.43; python_full_version >= '3.7.0'
pydot==2.0.0
pygments==2.17.2; python_version >= '3.7'
pyparsing==3.1.1
python-crontab==3.0.0
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1
pytz==2024.1
pyyaml==6.0.1
redis==5.0.1; python_version >= '3.7'
referencing==0.33.0; python_version >= '3.8'
requests==2.31.0; python_version >= '3.7'
rpds-py==0.17.1; python_version >= '3.8'
sentry-sdk[django]==1.40.3
setuptools==69.1.0; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
snowballstemmer==2.2.0
soupsieve==2.5; python_version >= '3.8'
sphinx==7.2.6; python_version >= '3.9'
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.8; python_version >= '3.9'
sphinxcontrib-devhelp==1.0.6; python_version >= '3.9'
sphinxcontrib-django==2.5
sphinxcontrib-htmlhelp==2.0.5; python_version >= '3.9'
sphinxcontrib-jquery==4.1; python_version >= '2.7'
sphinxcontrib-jsmath==1.0.1; python_version >= '3.5'
sphinxcontrib-qthelp==1.0.7; python_version >= '3.9'
sphinxcontrib-serializinghtml==1.1.10; python_version >= '3.9'
sqlparse==0.4.4; python_version >= '3.5'
tablib[html,ods,xls,xlsx,yaml]==3.5.0; python_version >= '3.8'
tomli==2.0.1; python_version < '3.11'
tornado==6.4; python_version >= '3.8'
types-pytz==2024.1.0.20240203; python_version >= '3.8'
types-pyyaml==6.0.12.12
typing-extensions==4.9.0; python_version >= '3.8'
tzdata==2024.1
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.2.0; python_version >= '3.6'
vine==5.1.0; python_version >= '3.6'
virtualenv==20.25.0; python_version >= '3.7'
wcwidth==0.2.13
xlrd==2.0.1
xlwt==1.3.0
